requirements.txt
pycryptodome==3.10.1
psutil==5.8.0
kaitaistruct==0.9